Added camera controls

+++ Source/Annex/Graphics/GameWindow.cs
@@ -6,8 +6,8 @@ namespace Annex.Graphics
     public class GameWindow : Singleton
     {
         public GraphicsContext Context { get; private set; }
-        public const float RESOLUTION_WIDTH = 1000;
-        public const float RESOLUTION_HEIGHT = 1000;
+        public const uint RESOLUTION_WIDTH = 1000;
+        public const uint RESOLUTION_HEIGHT = 1000;
 
         public GameWindow() {
             this.Context = new SfmlContext();
